fix: use exported constants in FVM/cgo errors
That way we can't get a mismatch.
