+++ rust/src/fvm/mod.rs
@@ -1,6 +1,5 @@
 /// cbindgen:ignore
 mod blockstore;
-/// cbindgen:ignore
 mod cgo;
 /// cbindgen:ignore
 mod externs;
